+++ Changelog.md
@@ -3,6 +3,7 @@ WhateverGreen Changelog
 #### v1.3.0
 - Fixed custom connector support for Radeon GPUs, thx @lwfitzgerald
 - Added `disable-gfx-submit` property to back `ngfxsubmit=0` boot argument
+- Added GuC firmware loading patch for latest SKL+ drivers
 
 #### v1.2.9
 - Added AMD Radeon VII to detected list